requirements.txt
streamlit==1.51.0
python-dotenv
langchain==1.0.3
langchain-groq==1.0.0
langchain-community==0.4.1
langchain-huggingface==1.0.0
qdrant-client==1.15.1